Repo renamed and files organized in directories.
