+++ medium/Bender - Episode 1.cpp
