+++ medium/Dwarfs standing on the shoulders of giants.cpp
