+++ medium/Scrabble.cpp
