+++ medium/Shadows of the Knight - Episode 1.cpp
