+++ medium/Stock Exchange Losses.cpp
