+++ medium/Teads Sponsored Contest.cpp
