+++ medium/The Last Crusade - Episode 1.cpp
