+++ medium/War.cpp
